backend/requirements.txt
django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.7.0
python-dotenv==1.0.1
boto3==1.37.13
openpyxl==3.1.5
Pillow==11.1.0
python-magic==0.4.27
django-storages==1.14.5
mistralai==1.5.1
requests==2.32.3
whitenoise[brotli]==6.9.0
gunicorn==22.0.0
uvicorn==0.29.0
